move interfaces to models folder
