+++ src/app/calculator/system-and-equipment-curve/equipment-curve/by-data-form/by-data-form.component.ts
@@ -2,9 +2,10 @@ import { Component, OnInit, Input } from '@angular/core';
 import { Settings } from '../../../../shared/models/settings';
 import { FormGroup, Validators, FormBuilder } from '@angular/forms';
 import { EquipmentCurveService } from '../equipment-curve.service';
-import { SystemAndEquipmentCurveService, ByDataInputs } from '../../system-and-equipment-curve.service';
+import { SystemAndEquipmentCurveService } from '../../system-and-equipment-curve.service';
 import { Subscription } from 'rxjs';
 import { CurveDataService } from '../../curve-data.service';
+import { ByDataInputs } from '../../../../shared/models/system-and-equipment-curve';
 
 @Component({
   selector: 'app-by-data-form',
