+++ src/app/calculator/system-and-equipment-curve/system-and-equipment-curve.service.ts
@@ -6,35 +6,40 @@ import { FanSystemCurveFormService } from './system-curve/fan-system-curve-form.
 import { PumpSystemCurveFormService } from './system-curve/pump-system-curve-form.service';
 import { PSAT } from '../../shared/models/psat';
 import { FSAT } from '../../shared/models/fans';
+import { ByDataInputs, ByEquationInputs, EquipmentInputs, PumpSystemCurveData, FanSystemCurveData } from '../../shared/models/system-and-equipment-curve';
 
 @Injectable()
 export class SystemAndEquipmentCurveService {
-  //persistent equipment curve data
+  //persistent data objects for calculator
   pumpByDataInputs: ByDataInputs;
   pumpByEquationInputs: ByEquationInputs;
   pumpEquipmentInputs: EquipmentInputs;
-
   fanByDataInputs: ByDataInputs;
   fanByEquationInputs: ByEquationInputs;
   fanEquipmentInputs: EquipmentInputs;
 
-  //behavior subjects
+  //behavior subjects for calculator state
   currentField: BehaviorSubject<string>;
-  pumpSystemCurveData: BehaviorSubject<PumpSystemCurveData>;
-  fanSystemCurveData: BehaviorSubject<FanSystemCurveData>;
   focusedCalculator: BehaviorSubject<string>;
-  byDataInputs: BehaviorSubject<ByDataInputs>;
-  equipmentInputs: BehaviorSubject<EquipmentInputs>;
-  byEquationInputs: BehaviorSubject<ByEquationInputs>;
   selectedEquipmentCurveFormView: BehaviorSubject<string>;
   equipmentCurveCollapsed: BehaviorSubject<string>;
   systemCurveCollapsed: BehaviorSubject<string>;
 
-  baselineEquipmentCurveDataPairs: BehaviorSubject<Array<{ x: number, y: number }>>;
-  modifiedEquipmentCurveDataPairs: BehaviorSubject<Array<{ x: number, y: number }>>;
-  systemCurveRegressionData: BehaviorSubject<Array<{ x: number, y: number, fluidPower: number }>>;
+  //form data behavior subjects
+  pumpSystemCurveData: BehaviorSubject<PumpSystemCurveData>;
+  fanSystemCurveData: BehaviorSubject<FanSystemCurveData>;
+  byDataInputs: BehaviorSubject<ByDataInputs>;
+  equipmentInputs: BehaviorSubject<EquipmentInputs>;
+  byEquationInputs: BehaviorSubject<ByEquationInputs>;
+
+  //calcuated data points
+  baselineEquipmentCurveDataPairs: Array<{ x: number, y: number }>;
+  modifiedEquipmentCurveDataPairs: Array<{ x: number, y: number }>;
+  systemCurveRegressionData: Array<{ x: number, y: number, fluidPower: number }>;
+
   //data points for system curve dropdown in assessment
   systemCurveDataPoints: Array<{ pointName: string, flowRate: number, yValue: number }>;
+  
   constructor(private equipmentCurveService: EquipmentCurveService,
     private fanSystemCurveFormService: FanSystemCurveFormService, private pumpSystemCurveFormService: PumpSystemCurveFormService) {
     this.currentField = new BehaviorSubject<string>('default');
@@ -47,67 +52,5 @@ export class SystemAndEquipmentCurveService {
     this.selectedEquipmentCurveFormView = new BehaviorSubject<string>('Equation');
     this.equipmentCurveCollapsed = new BehaviorSubject<string>('closed');
     this.systemCurveCollapsed = new BehaviorSubject<string>('closed');
-    this.baselineEquipmentCurveDataPairs = new BehaviorSubject(undefined);
-    this.modifiedEquipmentCurveDataPairs = new BehaviorSubject(undefined);
   }
-
-
-
-}
-
-export interface SystemAndEquipmentCurveData {
-  pumpSystemCurveData?: PumpSystemCurveData,
-  fanSystemCurveData?: FanSystemCurveData,
-  byEquationInputs?: ByEquationInputs,
-  byDataInputs?: ByDataInputs,
-  equipmentInputs?: EquipmentInputs,
-  equipmentCurveFormView?: string,
-  systemCurveDataPoints?: Array<{ pointName: string, flowRate: number, yValue: number }>
-
-}
-
-
-export interface PumpSystemCurveData {
-  specificGravity: number,
-  systemLossExponent: number,
-  pointOneFlowRate: number,
-  pointOneHead: number,
-  pointTwo: string,
-  pointTwoFlowRate: number,
-  pointTwoHead: number,
-}
-
-export interface FanSystemCurveData {
-  compressibilityFactor: number,
-  systemLossExponent: number,
-  pointOneFlowRate: number,
-  pointOnePressure: number,
-  pointTwo: string,
-  pointTwoFlowRate: number,
-  pointTwoPressure: number
-}
-
-
-export interface ByEquationInputs {
-  maxFlow: number,
-  equationOrder: number,
-  constant: number,
-  flow: number,
-  flowTwo: number,
-  flowThree: number,
-  flowFour: number,
-  flowFive: number,
-  flowSix: number
-}
-
-export interface EquipmentInputs {
-  measurementOption: number,
-  baselineMeasurement: number,
-  modificationMeasurementOption: number,
-  modifiedMeasurement: number
-}
-
-export interface ByDataInputs {
-  dataRows: Array<{ flow: number, yValue: number }>,
-  dataOrder: number
 }
